requirement.txt
aiohttp==3.8.4
aioice==0.7.6
aiortc==1.4.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
av==10.0.0
bidict==0.22.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
dnspython==2.3.0
frozenlist==1.3.3
google-crc32c==1.5.0
idna==3.4
multidict==6.0.4
netifaces==0.11.0
numpy==1.24.2
opencv-python==4.7.0.68
pycparser==2.21
pyee==9.0.4
pylibsrtp==0.8.0
pyOpenSSL==23.0.0
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.2
typing-extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
